Move pyDynamicCamMap to its own file
